--HG--
rename : editor/test/qa-functional/src/org/netbeans/test/editor/EditorKitsRegistryTest.java => ide.kit/test/qa-functional/src/org/netbeans/test/editor/EditorKitsRegistryTest.java
rename : ide.kit/test/whitelist.txt => ide.kit/test/qa-functional/data/whitelist.txt
files : apisupport.project/src/org/netbeans/modules/apisupport/project/Evaluator.java
files : apisupport.project/src/org/netbeans/modules/apisupport/project/NbModuleProject.java
files : apisupport.project/src/org/netbeans/modules/apisupport/project/ui/Bundle.properties
files : apisupport.project/src/org/netbeans/modules/apisupport/project/ui/ModuleActions.java
files : cnd.makeproject/src/org/netbeans/modules/cnd/makeproject/api/configurations/ServerConfiguration.java
files : cnd.remote/src/META-INF/services/org.netbeans.modules.cnd.api.remote.ServerList
files : core.ui/src/org/netbeans/core/ui/resources/PrintSettings.settings
files : debugger.jpda.ui/src/org/netbeans/modules/debugger/jpda/resources/root.png
files : debugger.jpda.ui/src/org/netbeans/modules/debugger/jpda/resources/sources.settings
files : debugger.jpda.ui/src/org/netbeans/modules/debugger/jpda/resources/sources.wstcgrp
files : debugger.jpda.ui/src/org/netbeans/modules/debugger/jpda/resources/sources.wstcref
files : debugger.jpda.ui/src/org/netbeans/modules/debugger/jpda/ui/actions/SourcesViewAction.java
files : debugger.jpda.ui/src/org/netbeans/modules/debugger/jpda/ui/views/Bundle.properties
files : editor/nbproject/project.xml
files : groovy.editor/src/org/netbeans/modules/groovy/editor/GroovyModuleInstall.java
files : groovy.editor/test/unit/src/org/netbeans/modules/groovy/editor/parser/GroovyVirtualSourveProviderTest.java
files : ide.kit/test/qa-functional/data/whitelist.txt
files : ide.kit/test/qa-functional/src/org/netbeans/test/editor/EditorKitsRegistryTest.java
files : j2ee.sun.appsrv81/src/org/netbeans/modules/j2ee/sun/ide/j2ee/mime-resolver.xml
files : nbbuild/build.xml
files : nbbuild/cluster.properties
files : nbbuild/licenses/TOM-SAWYER-SLA-6.0
files : nbbuild/templates/projectized.xml
files : nbjunit/apichanges.xml
files : nbjunit/manifest.mf
files : nbjunit/src/org/netbeans/junit/NbTestCase.java
files : nbjunit/src/org/netbeans/junit/NbTestSuite.java
files : spring.beans/src/org/netbeans/modules/spring/beans/completion/QueryProgress.java
files : spring.beans/test/unit/src/org/netbeans/modules/spring/beans/completion/CompletorTest.java
files : uml/src/org/netbeans/modules/uml/resources/license.html
files : websvc.rest.samples/CustomerDB/src/java/customerdb/converter/CustomerRefConverter.java
files : websvc.rest.samples/CustomerDB/src/java/customerdb/converter/DiscountCodeRefConverter.java
files : websvc.rest.samples/src/org/netbeans/modules/websvc/rest/samples/resources/projectIcon.gif
files : websvc.wsitmodelext/src/org/netbeans/modules/websvc/wsitmodelext/addressing/impl/Addressing10Attribute.java
files : websvc.wsitmodelext/src/org/netbeans/modules/websvc/wsitmodelext/addressing/impl/AddressingAttribute.java

+++ nbjunit/src/org/netbeans/junit/NbTestCase.java
@@ -126,6 +126,18 @@ public abstract class NbTestCase extends TestCase implements NbTest {
      * @return true if the test can run
      */
     public boolean canRun() {
+        if (NbTestSuite.ignoreRandomFailures()) {
+            if (getClass().isAnnotationPresent(RandomlyFails.class)) {
+                return false;
+            }
+            try {
+                if (getClass().getMethod(getName()).isAnnotationPresent(RandomlyFails.class)) {
+                    return false;
+                }
+            } catch (NoSuchMethodException x) {
+                // Specially named methods; let it pass.
+            }
+        }
         if (null == filter) {
             //System.out.println("NBTestCase.canRun(): filter == null name=" + name ());
             return true; // no filter was aplied
@@ -753,7 +765,7 @@ public abstract class NbTestCase extends TestCase implements NbTest {
     
     // private method for deleting every subfiles/subdirectories of a file object
     static void deleteSubFiles(File file) throws IOException {
-        if (file.isDirectory()) {
+        if (file.isDirectory() && file.exists()) {
             File files[] = file.listFiles();
             for (int i = 0; i < files.length; i++) {
                 deleteFile(files[i]);
